路由架构调整（未完成）
