+++ library/think/route/dispatch/Controller.php
@@ -17,9 +17,9 @@ class Controller extends Dispatch
 {
     public function run()
     {
-        // 执行控制器操作
+        // 执行控制器的操作方法
         $vars = array_merge($this->app['request']->param(), $this->param);
-        return $this->action($this->action, $vars, $this->app->config('app.url_controller_layer'), $this->app->config('app.controller_suffix'));
+        return $this->app->action($this->action, $vars, $this->app->config('app.url_controller_layer'), $this->app->config('app.controller_suffix'));
     }
 
 }
